test(cypress tests): improve Cypress test stability for low-resource environments
