+++ cypress/component/Modal-test-1.cy.tsx
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 import { useState } from 'react'
-import { Tooltip, Modal, Button, CloseButton, View } from '@instructure/ui'
+import { Tooltip, Modal, Button, CloseButton } from '@instructure/ui'
 import 'cypress-real-events'
 
 import '../support/component'
@@ -41,70 +41,6 @@ describe('<Modal/>', () => {
       </Modal>
     )
     cy.root().should('contain', 'Modal body text')
-  })
-
-  it('should call onDismiss prop when Esc key pressed by default', () => {
-    const onDismissSpy = cy.spy()
-    cy.mount(
-      <Modal
-        open={true}
-        onDismiss={onDismissSpy}
-        label="Modal Dialog"
-        shouldReturnFocus={false}
-      >
-        <p>Modal body text</p>
-      </Modal>
-    )
-    cy.root().should('contain', 'Modal body text')
-
-    cy.get('body').trigger('keydown', { keyCode: 27 })
-    cy.get('body').trigger('keyup', { keyCode: 27 })
-    cy.wrap(onDismissSpy).should('have.been.calledOnce')
-  })
-
-  it('should not call stale callbacks', () => {
-    const handleDismissSpy = cy.spy()
-    interface ExampleProps {
-      handleDismiss: (value: number) => void
-    }
-
-    function Example({ handleDismiss }: ExampleProps) {
-      const [value, setValue] = useState(0)
-
-      function onButtonClick() {
-        setValue(value + 1)
-      }
-
-      return (
-        <View>
-          <Modal
-            label="Modal"
-            open
-            onDismiss={() => {
-              handleDismiss(value)
-            }}
-          >
-            <Modal.Body>
-              <p>Modal body text</p>
-              <div id="value-indicator">{value}</div>
-              <button id="increment-btn" onClick={onButtonClick}>
-                Increment Button
-              </button>
-            </Modal.Body>
-          </Modal>
-        </View>
-      )
-    }
-    cy.mount(<Example handleDismiss={handleDismissSpy} />)
-
-    cy.root().should('contain', 'Modal body text')
-
-    cy.get('#increment-btn').click().wait(100)
-    cy.get('#value-indicator').should('contain', '1')
-    cy.wrap(handleDismissSpy).should('not.have.been.called')
-
-    cy.get('body').click(0, 0)
-    cy.wrap(handleDismissSpy).should('have.been.calledOnceWith', 1)
   })
 
   it('should not close when button is clicked to rerender content', () => {
