+++ cypress/component/Popover.cy.tsx
@@ -121,44 +121,45 @@ describe('<Popover/>', () => {
       })
   })
 
-  it('should hide content when clicked outside content by default', () => {
-    const onHideContent = cy.spy()
-    cy.mount(
-      <div id="main">
-        <div>
-          <button>Outer</button>
-        </div>
-        <div>
-          <Popover
-            on="click"
-            onHideContent={onHideContent}
-            renderTrigger={<button>Trigger</button>}
-          >
-            <h2>Popover content</h2>
-            <button>focus me</button>
-          </Popover>
-        </div>
-      </div>
-    )
-    cy.contains('button', 'Trigger')
-      .realClick()
-      .then(() => {
-        cy.contains('Popover content').should('be.visible')
-      })
-      .then(() => {
-        cy.contains('button', 'Outer').click(0, 0)
-      })
-      .then(() => {
-        cy.contains('Popover content').should('not.exist')
-        cy.wrap(onHideContent)
-          .should('have.been.calledOnce')
-          .then((spy) => {
-            cy.wrap(spy)
-              .its('lastCall.args')
-              .should('deep.include', { documentClick: true })
-          })
-      })
-  })
+  // TODO convert to e2e regression
+  // it('should hide content when clicked outside content by default', () => {
+  //   const onHideContent = cy.spy()
+  //   cy.mount(
+  //     <div id="main">
+  //       <div>
+  //         <button>Outer</button>
+  //       </div>
+  //       <div>
+  //         <Popover
+  //           on="click"
+  //           onHideContent={onHideContent}
+  //           renderTrigger={<button>Trigger</button>}
+  //         >
+  //           <h2>Popover content</h2>
+  //           <button>focus me</button>
+  //         </Popover>
+  //       </div>
+  //     </div>
+  //   )
+  //   cy.contains('button', 'Trigger')
+  //     .realClick()
+  //     .then(() => {
+  //       cy.contains('Popover content').should('be.visible')
+  //     })
+  //     .then(() => {
+  //       cy.contains('button', 'Outer').click(0, 0)
+  //     })
+  //     .then(() => {
+  //       cy.contains('Popover content').should('not.exist')
+  //       cy.wrap(onHideContent)
+  //         .should('have.been.calledOnce')
+  //         .then((spy) => {
+  //           cy.wrap(spy)
+  //             .its('lastCall.args')
+  //             .should('deep.include', { documentClick: true })
+  //         })
+  //     })
+  // })
 
   it('should move focus into the content when the trigger is blurred', () => {
     const onHideContent = cy.spy()
@@ -185,17 +186,18 @@ describe('<Popover/>', () => {
 
     cy.contains('button', 'trigger btn initial focus me')
       .focus()
+      .wait(200)
       .then(() => {
         cy.contains('focus me after trigger').should('not.be.focused')
       })
       .then(() => {
-        cy.realPress('Tab')
+        cy.realPress('Tab').wait(200)
       })
       .then(() => {
         cy.contains('focus me after trigger').should('be.focused')
       })
       .then(() => {
-        cy.realPress('Tab')
+        cy.realPress('Tab').wait(200)
       })
       .then(() => {
         cy.wrap(onHideContent).should('have.been.calledOnce')
