+++ cypress/component/TextArea.cy.tsx
@@ -42,8 +42,8 @@ it('should resize if autoGrow is true', () => {
       )
       .trigger('input')
 
-    cy.get('textarea').then((input) => {
-      const resizedHeight = parseInt(getComputedStyle(input[0]).height, 10)
+    cy.get('textarea').should(($input) => {
+      const resizedHeight = parseInt(getComputedStyle($input[0]).height, 10)
       expect(resizedHeight).to.be.above(initialHeight)
     })
